// src/app/models/studentModel.ts
export interface StudentModel {
  id: string;
  firstName: string;
  lastName: string;
  paymentMethod?: string;
  paymentHistory?: Date;
  dataSource: string;
  no: string;
  isReg: string;
  interestedEducation: string;
  staff: string;
  careerCounselor: string;
  collection?: number;
  report: string;
  email: string;
  tc: string;
  location: string;
  collectionAmount: string;
  advancePay: string;
  educationHistory: Date;
  regHistory: Date;
  numberBills: string;
}